check if file exists

+++ grab.js
@@ -16,6 +16,19 @@ folder = system.args[2];
 nThreads = parseInt(system.args[4]);
 
 phantom.injectJs("md5.js");
+var nextStep = function(index) {
+    index += nThreads;
+    if (index < url.length) {
+        savePage(index);
+    } else {
+        console.log("end of thread");
+        finishedThreads += 1;
+        if (finishedThreads == nThreads) {
+            console.log("end.");
+            phantom.exit();
+        }
+    }
+}
 var savePage = function(index) {
     address = url[index];
     var page = webpage.create();
@@ -28,55 +41,40 @@ var savePage = function(index) {
     };
     page.onError = undefined;
     console.log("pre-open url: " + address + " index: " + index);
-    page.open(address, function(status) {
-        //console.log("loaded? url: " + address);
-        if (status !== 'success') {
-            console.log('FAIL to load the address. status: ' + statuis);
-        } else {
+    var file_name = folder + "/" + CryptoJS.MD5(address) + ".html";
+    if (fs.exists(file_name)) {
+        console.log("file " + file_name + " exist, not retrieving");
+        nextStep(index);
+    } else {
+        //console.log("not exist, retrieving");
+        page.open(address, function(status) {
+            //console.log("loaded? url: " + address);
+            if (status !== 'success') {
+                console.log('FAIL to load the address. status: ' + statuis);
+            } else {
 
-            var file_name = folder + "/" + CryptoJS.MD5(address) + ".html";
-            console.log(file_name);
-            fs.write(file_name, page.content, 'w');
+                console.log(file_name);
+                fs.write(file_name, page.content, 'w');
 
-        }
-        index += nThreads;
-        if (index < url.length) {
-            savePage(index);
-        } else {
-            console.log("end of thread");
-            finishedThreads += 1;
-            if (finishedThreads == nThreads) {
-                console.log("end.");
-                phantom.exit();
             }
-        }
+            nextStep(index);
 
-        //phantom.exit();
-    });
+            //phantom.exit();
+        });
+    }
+
 
 }
 
 
 
 var urls = '',
-    f = null,
-    lines = null,
     eol = system.os.name == 'windows' ? "\r\n" : "\n";
-
-try {
-    f = fs.open(list_urls, "r");
-    urls = f.read();
-} catch (e) {
-    console.log(e);
-}
-
-if (f) {
-    f.close();
-}
+urls = fs.read(list_urls);
 
 if (urls) {
     url = urls.split(eol);
-    /*	for (var i = 0, len = url.length; i < len; i++) {
+    /*for (var i = 0, len = url.length; i < len; i++) {
 		console.log(url[i]);
 		}*/
 }
